+++ src/Spago/Build.hs
@@ -278,6 +278,12 @@ script modulePath tag packageDeps opts = do
   absoluteModulePath <- fmap Text.pack (makeAbsolute (Text.unpack modulePath))
   currentDir <- Turtle.pwd
 
+  compilerVer <- Purs.pursVersion >>= \case
+    Left t ->
+      pure $ Text.unpack t
+    Right v ->
+      pure $ show v
+
   -- This is the part where we make sure that the script reuses the same folder
   -- if run with the same options more than once. We do that by making a folder
   -- in the system temp directory, and naming it with the hash of the script
@@ -285,7 +291,7 @@ script modulePath tag packageDeps opts = do
   let sha256 :: String -> String
       sha256 = show . (Hash.hash :: ByteString -> Hash.Digest Hash.SHA256) . Turtle.fromString
   systemTemp <- liftIO $ Temp.getCanonicalTemporaryDirectory
-  let stableHash = sha256 (Text.unpack absoluteModulePath <> show tag <> show packageDeps <> show opts)
+  let stableHash = sha256 (compilerVer <> Text.unpack absoluteModulePath <> show tag <> show packageDeps <> show opts)
   let scriptDirPath = Turtle.decodeString (systemTemp </> "spago-script-tmp-" <> stableHash)
   logDebug $ "Found a system temp directory: " <> displayShow systemTemp